# failiverCheck/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
...
        "/system_calcs/": {
            "get": {
                "description": "get system cacl in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System caclculation"
                ],
                "summary": "Show an system calc list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name search by q",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2006-02-01",
                        "name": "date_formed_start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "2006-02-01",
                        "name": "date_formed_end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SystemCalculationInfoDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            }
        },
        "/system_calcs/my_bucket": {
            "get": {
                "description": "get user bucket system cac",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System caclculation"
                ],
                "summary": "Show user bucket",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CurrentUserBucketDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            }
        },
        "/system_calcs/{id}": {
            "get": {
                "description": "get system cacl by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System caclculation"
                ],
                "summary": "Show an system calc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "System Cacl ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SystemCalculationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "update system cacl by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System caclculation"
                ],
                "summary": "Update system calc by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "System Cacl ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update schema",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateSystemCalcFields"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SystemCalculationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete system calc by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System caclculation"
                ],
                "summary": "Delete system calc by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "System Cacl ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/schemas.OKResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            }
        },
        "/system_calcs/{id}/status": {
            "put": {
                "description": "Moderate system calc, update status by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System caclculation"
                ],
                "summary": "Moderate system calc",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "System Cacl ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SystemCalculationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            }
        },
        "/system_calcs/{id}/status_formed": {
            "put": {
                "description": "Update system calc status to FORMED by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System caclculation"
                ],
                "summary": "Update system calc status to FORMED by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "System Cacl ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SystemCalculationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ComponentInSystemCalcDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.ComponentsToSystemCalcDTO": {
            "type": "object",
            "properties": {
                "component": {
                    "$ref": "#/definitions/dto.ComponentInSystemCalcDTO"
                },
                "replication_count": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CurrentUserBucketDTO": {
            "type": "object",
            "properties": {
                "components_count": {
                    "type": "integer"
                },
                "sys_calculation_id": {
                    "type": "integer"
                }
            }
        },
        "dto.SystemCalculationDTO": {
            "type": "object",
            "properties": {
                "available_calculation": {
                    "type": "number"
                },
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ComponentsToSystemCalcDTO"
                    }
                },
                "date_accepted": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_formed": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "system_name": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "dto.SystemCalculationInfoDTO": {
            "type": "object",
            "properties": {
                "available_calculation": {
                    "type": "number"
                },
                "date_accepted": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_formed": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "system_name": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "schemas.Error": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "schemas.OKResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "schemas.UpdateSystemCalcFields": {
            "type": "object",
            "required": [
                "system_name"
            ],
            "properties": {
                "system_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}